Zierman - Created Request classes for each request type (#19)
* Use with open to ensure f is closed even if an error is encountered.

* Update docstrings and type hints

* added doc folder to .gitignore

generated documentation needs not be included in the repository at this time.

* Added script to generate documentation

* Changed the log method to use label argument if provided.

* Added was_success class method to Status class.

The method returns a status string constant depending on the was_success value.

* Work on message_items and util.const

- created subclasses for each type of request
- cleaned things up a bit
- added util.const module

* Fixed _unimplemented function

* added needed imports

* Addressing Code Inspection Stuff

* Add inspectionProfiles to .gitignore

* Worked on tests for util.const.py

* setup quick db check on startup

- Added a check when initializing a DB object to check that it can actually connect to DB (before any requests come in)
- Reworking the mutli-threading and multi-processing.
    - I replaced the Processes with Threads so that we can keep memory usage lower.
    - I added timeouts to ensure we never block indefinitely
    - I added a system for communicating when the service is dying.

* add info log at end of program

* addressed inspection warnings

* addressed inspection warnings

* bug fixes

* bug fixes

* bug fixes

* bug fixes

* bug fixes

* bug hunt

* bug hunt

* bug hunt

* bug hunt

* bug hunt

* bug hunt

* Made the git_pull.sh check if root.

* removed logging of lower layer db methods.

We can add them back easy enough if we need to troubleshoot something.

* Bug hunting

* tweaking how the pull request works

* Potential solution to why service fails to start

* bug hunt

* bug hunt

* make more status line print

* checking pyscript

* checking pyscript

* Revert "checking pyscript"

This reverts commit 8df12b3e

* checking pyscript

* checking pyscript

* checking pyscript

* checking pyscript

* checking pyscript

* checking pyscript

* checking pyscript

* checking pyscript

* checking pyscript

* checking pyscript

* checking pyscript

* checking pyscript

* bug hunt

* bug fix

* bug hunt

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* work on tmp.py

* I'm clearing the log on reset

We may want to change this keep a rotation of so many logs (or do that from within the service) but for now I want to make the log as easy as possible to read when restarting the service.

* Added the log output back in.

* Undo something I tried.

It doesn't do much to load project path from a json stored within the project's directory because you need to be able to find it to use it, which means you already found the project directory.

I was having issues because the service was starting with a working directory of the file system root. I since figured out why I had the issue and corrected it.

* Introduced inline subtest annotation

* renamed ConstContainer class and work on jargit

* made abstract methods raise NotImplementedError

* added depreciated decorator to global_logger.py

* removed garbage comment

* changed depreciated to log on call.

* Used the new depreciated decorator

Used to mark send_request_data and send_account_creation_status as depreciated and removed the now unnecessary log from within the methods.

* revert controller, listener, and process_request.py

* update docstring

* address inspection issues

* spelling fix

* addressing inspection issues

* added timeout of 10 seconds

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* debugging

* done debugging

+++ git_pull.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
+if [ $(id -u) -ne 0 ]
+  then echo "Unauthorized"
+  exit
+fi
 systemctl stop jar_logon.service
+echo '' > ./logs/logon_server.log # clear log
 git stash
 git fetch
 git pull
@@ -9,5 +14,6 @@ chmod +x ./controller.py
 chmod 0554 -R *
 chmod 0664 ./logs/logon_server.log
 chmod 0440 ./params.json
-systemctl restart jar_logon.service
+systemctl start jar_logon.service
+systemctl status -n 100 jar_logon.service
 tail -n 100 -f ./logs/logon_server.log
